Improve API shape for registering sys web adapters in NetFx
