+++ samples/RemoteAuth/Bearer/RemoteBearer/Global.asax.cs
@@ -11,7 +11,7 @@ namespace RemoteOAuth
             GlobalConfiguration.Configure(WebApiConfig.Register);
             FilterConfig.RegisterGlobalFilters(GlobalFilters.Filters);
 
-            this.AddSystemWebAdapters()
+            SystemWebAdapterConfiguration.AddSystemWebAdapters(this)
                 .AddRemoteApp(options =>
                 {
                     // A real application would not hard code this, but load it
